backend/requirements.txt
# FastAPI and web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Pydantic for data validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# LDAP integration
ldap3==2.9.1

# RAG and ML
sentence-transformers==2.3.1
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.0

# Environment and utilities
python-dotenv==1.0.0
